Multiple USB update
This update allows for the installation on multiple USB devices without needing to re-enter the keys.

To select the multiple update option, you will need to enter Y when asked if you want to setup another USB key. Make sure to insert your USB drive, and unlocked it, before pressing enter to note that you have named and inserted and unlocked the drive.